5.3_adding_lambda_authorizers/requirements.txt
ecdsa==0.14.1
pyasn1==0.4.8
python-jose==3.1.0
rsa==4.0
six==1.13.0
